src/requirements.txt
oauth2client==4.1.3
google-api-python-client==2.64.0
bs4
requests
selenium
pydantic==1.9.1
pandas
pyinstaller